Extract deepEqual to own file
